backend/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
python-multipart==0.0.22
pydantic-settings==2.13.1
langchain==1.2.10
langchain-core==1.2.14
langchain-openai==1.1.10
langchain-huggingface==1.2.0
langchain-pinecone==0.2.13
langchain-community==0.4.1
langchain-text-splitters==1.1.1
pypdf==6.7.2
pinecone-client==6.0.0
openai-whisper==20250625
edge-tts==7.2.7
soundfile==0.13.1
numpy==2.4.2
requests==2.32.5
